Adding color chooser preferences
We need the ability to select which color choosers are available/allowed
I have therefore started adding a Color chooser preferences page to the
Window builder preferces

Removing unnecessary preference options

The preference that allows the user to order the tabs in the color
property editor is deemed unnecessary. It has therefore been removed.

Updating ColorChooser preference to disregard changes that were made and
not applied. I.e. setting the checkboxes to the previously saved state

Removing unnecessary constants from the IColorChooserPreferenceConstants

Updating ModelMessages to contain the new string values used on the
preferences

Updating IColorChooserPreferenceConstants to obtain the names of the
various options from ModelMessages instead of hardcoding the string
values