Adding a popup menu to allow the user to specify the layout of the icons on the Palette
A popup menu has been added to allow the user to quickly specify the
layout of the icons on the palette. Columns, List, Icons only or icons
with details. Furthermore, when selecting settings.The user can also
specify the font used for the category descriptions and icon
descriptions. However when changing the font the editor needs to be
closed and reopened again. Within these settings the user can also
specify the layout as well as make modifications to the number of
columns to be used for the layout